feature: send cred_v on message_2
